+++ deeplearning4j/deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/conf/graph/PoolHelperVertex.java
@@ -131,12 +131,11 @@ public class PoolHelperVertex extends GraphVertex {
             //CNN inputs... also check that the channels, width and heights match:
             InputType.InputTypeConvolutional firstConv = (InputType.InputTypeConvolutional) first;
 
-            // FIXME: int cast
-            val fd = (int) firstConv.getChannels();
-            val fw = (int) firstConv.getWidth();
-            val fh = (int) firstConv.getHeight();
+            val fd = firstConv.getChannels();
+            val fw = firstConv.getWidth();
+            val fh = firstConv.getHeight();
 
-            int depthSum = fd;
+            long depthSum = fd;
 
             for (int i = 1; i < vertexInputs.length; i++) {
                 if (vertexInputs[i].getType() != InputType.Type.CNN) {
@@ -148,10 +147,9 @@ public class PoolHelperVertex extends GraphVertex {
 
                 InputType.InputTypeConvolutional otherConv = (InputType.InputTypeConvolutional) vertexInputs[i];
 
-                // FIXME: int cast
-                int od = (int) otherConv.getChannels();
-                int ow = (int) otherConv.getWidth();
-                int oh = (int) otherConv.getHeight();
+                long od = otherConv.getChannels();
+                long ow = otherConv.getWidth();
+                long oh = otherConv.getHeight();
 
                 if (fw != ow || fh != oh) {
                     throw new InvalidInputTypeException(
