+++ deeplearning4j/deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/conf/graph/UnstackVertex.java
@@ -150,12 +150,11 @@ public class UnstackVertex extends GraphVertex {
             //CNN inputs... also check that the channels, width and heights match:
             InputType.InputTypeConvolutional firstConv = (InputType.InputTypeConvolutional) first;
 
-            // FIXME: int cast
-            val fd = (int) firstConv.getChannels();
-            val fw = (int) firstConv.getWidth();
-            val fh = (int) firstConv.getHeight();
+            val fd = firstConv.getChannels();
+            val fw = firstConv.getWidth();
+            val fh = firstConv.getHeight();
 
-            int depthSum = fd;
+            long depthSum = fd;
 
             for (int i = 1; i < vertexInputs.length; i++) {
                 if (vertexInputs[i].getType() != InputType.Type.CNN) {
@@ -167,10 +166,9 @@ public class UnstackVertex extends GraphVertex {
 
                 InputType.InputTypeConvolutional otherConv = (InputType.InputTypeConvolutional) vertexInputs[i];
 
-                // FIXME: int cast
-                val od = (int) otherConv.getChannels();
-                val ow = (int) otherConv.getWidth();
-                val oh = (int) otherConv.getHeight();
+                val od = otherConv.getChannels();
+                val ow = otherConv.getWidth();
+                val oh = otherConv.getHeight();
 
                 if (fw != ow || fh != oh) {
                     throw new InvalidInputTypeException(
