+++ deeplearning4j/deeplearning4j-scaleout/spark/dl4j-spark/src/main/java/org/deeplearning4j/spark/api/worker/ExecuteWorkerMultiDataSetFlatMap.java
@@ -84,9 +84,8 @@ public class ExecuteWorkerMultiDataSetFlatMap<R extends TrainingResult> implemen
                     s.logNextDataSetBefore();
                 MultiDataSet next = batchedIterator.next();
 
-                // FIXME: int cast
                 if (stats)
-                    s.logNextDataSetAfter((int) next.getFeatures(0).size(0));
+                    s.logNextDataSetAfter(next.getFeatures(0).size(0));
 
                 if (stats) {
                     s.logProcessMinibatchBefore();
